refactor: platform/http をフラット化し httpclient.New に改名
- platform/http/{handler,middleware} を platform/ 直下へ移動し中間階層を削除
- httpclient.NewHTTPClient を httpclient.New に改名（パッケージ名と型名の重複解消）
- symbollist/README.md のディレクトリツリーで # コメント位置を全行揃える
- CLAUDE.md の platform ディレクトリツリーを実態に合わせて更新
