+++ internal/app/di/market.go
@@ -9,6 +9,6 @@ import (
 // NewMarket はHTTPクライアント付きの完全に設定されたTwelveDataMarketを生成します。
 func NewMarket() *twelvedata.TwelveDataMarket {
 	cfg := twelvedata.LoadConfig()
-	httpClient := httpclient.NewHTTPClient(cfg.Timeout)
+	httpClient := httpclient.New(cfg.Timeout)
 	return twelvedata.NewTwelveDataMarket(cfg, httpClient)
 }
