+++ internal/platform/handler/health_handler.go
