+++ internal/platform/handler/health_handler_test.go
