+++ internal/platform/middleware/security_headers.go
